Add files via upload
This is the final working main.cpp file for a program that calculates the number of street lights required along an Arc and their positions. 
Input: radius of arc, distance between the ends of the arc and the max spacing allowed between lights. 
Returns: Distance from the first Arc and the angle it makes with the cord

Uploaded by Adam Claxton